Detect_edits2
I had to modify the code in event_line and mean_and_trend as I found some errors after I changed detect.
Note that the mean_and_trend function still needs the Stats toolbox to function, as you use fitlm. We would have to change the function to polyfit to get rid of the toolbox dependency.